+++ website/src/pages/apps/ToolsPage.vue
@@ -46,7 +46,7 @@ const filteredCategories = computed(() => TOOL_BUTTONS_BY_CATEGORY.map(v => {
                 <SearchBar v-model="search" placeholder="Search tool" :debounce="300"/>
             </div>
             <template v-if="menuStore.pinnedTools.length > 0">
-                <h6 class="q-mt-xl">Pinned</h6>
+                <h6 class="q-mt-xl">Pinned tools</h6>
                 <q-separator class="q-mt-sm q-mb-md"/>
                 <div class="row justify-start items-stretch gap-md">
                     <ToolCard v-for="button in menuStore.pinnedTools" :key="button.name" :button="button"></ToolCard>